+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - Added early stopping window to AdaBoost
     - Gaussian MLE now has automatic and adaptive threshold
     - Loda now has automatic and adaptive threshold
+    - Variance Threshold Filter now selects top k features
     - Added params method to Estimator and Embedder interface
     - t-SNE now compatible with categorical distance kernels
     - Grid Search implements the Wrapper interface